getGlobalTime in ARICHSimHit was modified to return float (instead double).
git-svn-id: https://belle2.cc.kek.jp/svn/trunk/software@13922 b38c6c8f-bee8-465d-8448-9abe6915d496

+++ arich/dataobjects/include/ARICHSimHit.h
@@ -64,7 +64,7 @@ namespace Belle2 {
     TVector2 getLocalPosition() const { TVector2 vec(m_x, m_y); return vec; }
 
     //! Get global time of hit
-    double getGlobalTime() const { return m_globalTime; }
+    float getGlobalTime() const { return m_globalTime; }
 
     //! Get detected photon energy
     double getEnergy() const { return m_energy; }
